refactor(NA) - Separate package structure for each client factory
This allows a project to include both http and udp clients and decide which one to use at runtime.
